move plot to separate utils file
